requirements.txt
aiofiles==25.1.0
fastapi==0.128.0
Jinja2==3.1.6
openai==2.15.0
python-dotenv==1.2.1
python-multipart==0.0.21
uvicorn==0.40.0
websockets==16.0
